[DMS-274] Remove savepoints for backend reference validation (#208)
* [DMS-274] Remove savepoints for backend reference validation

* [DMS-274] remove experimental full upsert proc

* [DMS-274] add proc to project, fix dbup incompatability

* make npgsql happy

* typo

+++ src/backend/EdFi.DataManagementService.Backend/PartitionKey.cs
@@ -8,4 +8,4 @@ namespace EdFi.DataManagementService.Backend;
 /// <summary>
 /// Database integer partition key as a Branded type
 /// </summary>
-public record PartitionKey(int Value);
+public record PartitionKey(short Value);
